test: isolate default baseline output coverage
